refactor: adds separate folder to sort donations
